Historial en Supabase (Postgres) y deduplicación de alertas
- migrations/001_init.sql: assets, candles, signals, alerts, trades con RLS activo
- core/db.py: PostgresRepository (psycopg) y NullRepository sin DATABASE_URL
- jobs/migrate.py aplica migraciones pendientes
- Señal única por vela: si el job corre dos veces, no repite la alerta
- Si la DB falla, la alerta se envía igual y se reporta el error
- Nueva dependencia: psycopg[binary] (3 paquetes vs 31 del cliente supabase)
- docs/supabase.md y tests contra Postgres real (con TEST_DATABASE_URL)

+++ core/strategies/base.py
@@ -30,7 +30,7 @@ class BaseStrategy(ABC):
             asset=asset,
             strategy=self.name,
             direction=direction,
-            strength=round(min(max(strength, 0.0), 1.0), 3),
+            strength=float(round(min(max(strength, 0.0), 1.0), 3)),
             reason=reason,
             ts=candles.index[-1].to_pydatetime(),
         )
